copy from aokp script update

+++ cp_from_aokp.sh
@@ -1,17 +1,40 @@
 #!/bin/sh
-
-mkdir device/samsung/galaxys2
+if [ ! -d device/samsung/galaxys2 ];
+then
+   echo "Make directory device/samsung/galaxys2"
+   mkdir device/samsung/galaxys2
+fi
+echo "Copy directory device/samsung/galaxys2"
 cp -R ../aokp/device/samsung/galaxys2/* device/samsung/galaxys2
 
-mkdir vendor/samsung/galaxys2
-mkdir vendor/samsung/galaxys2-common
+if [ ! -d vendor/samsung/galaxys2 ];
+then
+   echo "Make directory vendor/samsung/galaxys2"
+   mkdir vendor/samsung/galaxys2
+fi
+if [ ! -d vendor/samsung/galaxys2-common ];
+then
+   echo "Make directory vendor/samsung/galaxys2-common"
+   mkdir vendor/samsung/galaxys2-common
+fi
+echo "Copy directory aokp/vendor/samsung/galaxys2"
 cp -R ../aokp/vendor/samsung/galaxys2/* vendor/samsung/galaxys2
+echo "Copy directory vendor/samsung/galaxys2-common"
 cp -R ../aokp/vendor/samsung/galaxys2-common/* vendor/samsung/galaxys2-common
 
-mkdir packages/apps/Settings
-cp -R ../aokp/packages/apps/Settings/* packages/apps/Settingsn
-rm packages/apps/Settings/.git
-
-mkdir kernel/samsung/smdk4210/arch/arm/configs
+if [ ! -d packages/apps/Settings ];
+then
+   mkdir packages/apps/Settings
+fi
+cp -R ../aokp/packages/apps/Settings/* packages/apps/Settings
+if [ -d packages/apps/Settings/.git ];
+then
+   rm packages/apps/Settings/.git
+fi
+if [ ! -d kernel/samsung/smdk4210/arch/arm/configs ];
+then
+   echo "LOOKS LIKE YOU ARE MISSING YOUR KERNEL FOLDER FROM CM9"
+   mkdir kernel/samsung/smdk4210/arch/arm/configs
+fi
 cp ../android/system/kernel/samsung/smdk4210/arch/arm/configs/cyanogenmod_i777_defconfig kernel/samsung/smdk4210/arch/arm/configs/cyanogenmod_i777_defconfig
